関数の変更
全てスコアのクラスに記入した
データ型の変更

+++ competition/Scenes/UI/Score.h
@@ -90,7 +90,7 @@ public:
 	static int GetPlayerLevel();
 	static int GetPlayerPower();
 	static int GetPlayerSpeed();
-	static float GetPlayerCoolTime();
+	static int GetPlayerCoolTime();
 	static int GetPlayerHitRange();
 	static bool GetPlayerThreeWay();
 	static int GetPlayerDrone();
